Added exceptions validator

+++ src/main/java/com/matera/cursoferias/petstore/business/ClienteBusiness.java
@@ -6,15 +6,20 @@ import java.util.List;
 import org.springframework.stereotype.Component;
 
 import com.matera.cursoferias.petstore.entity.Cliente;
+import com.matera.cursoferias.petstore.entity.Pet;
+import com.matera.cursoferias.petstore.exception.ResourceBadRequestException;
+import com.matera.cursoferias.petstore.exception.ResourceNotFoundException;
 import com.matera.cursoferias.petstore.repository.ClienteRepository;
 
 @Component
 public class ClienteBusiness implements ClienteBusinessInterface {
 
 	private ClienteRepository clienteRepository;
+	private PetBusiness petBusiness;
 	
-	public ClienteBusiness(ClienteRepository clienteRepository) {
+	public ClienteBusiness(ClienteRepository clienteRepository, PetBusiness petBusiness) {
 		this.clienteRepository = clienteRepository;
+		this.petBusiness = petBusiness;
 	}
 
 	@Override
@@ -33,11 +38,25 @@ public class ClienteBusiness implements ClienteBusinessInterface {
 
 	@Override
 	public Cliente findById(Long id) {
-		return clienteRepository.findById(id).orElse(null);
+		Cliente cliente = clienteRepository.findById(id).orElse(null);
+		
+		if (cliente == null) {
+			throw new ResourceNotFoundException(String.format("Cliente %d não encontrado!", id));
+		}
+		
+		return cliente;
 	}
 
 	@Override
 	public void deleteById(Long id) {
+		findById(id);
+		
+		List<Pet> pets = petBusiness.findByCliente_Id(id);
+		
+		if (!pets.isEmpty()) {
+			throw new ResourceBadRequestException(String.format("Cliente %d não pode ser excluído pois possuí %d pets", id, pets.size())); 
+		}
+		
 		clienteRepository.deleteById(id);
 	}
